+++ viadot/flows/sharepoint_to_adls.py
@@ -27,10 +27,10 @@ json_to_adls_task = AzureDataLakeUpload()
 class SharepointToADLS(Flow):
     def __init__(
         self,
-        name: str = None,
+        name: str,
+        url_to_file: str = None,
         nrows_to_df: int = None,
         path_to_file: str = None,
-        url_to_file: str = None,
         sheet_number: int = None,
         validate_excel_file: bool = False,
         output_file_extension: str = ".csv",
@@ -38,6 +38,7 @@ class SharepointToADLS(Flow):
         adls_dir_path: str = None,
         adls_file_name: str = None,
         adls_sp_credentials_secret: str = None,
+        overwrite_adls: bool = False,
         if_empty: str = "warn",
         if_exists: str = "replace",
         *args: List[any],
@@ -47,11 +48,11 @@ class SharepointToADLS(Flow):
         Flow for downloading Excel file from Sharepoint then uploading it to Azure Data Lake.
 
         Args:
-            name (str, optional): The name of the flow. Defaults to None.
+            name (str): The name of the flow.
+            url_to_file (str, optional): Link to a file on Sharepoint. Defaults to None.
+                        (e.g : https://{tenant_name}.sharepoint.com/sites/{folder}/Shared%20Documents/Dashboard/file).
             nrows_to_df (int, optional): Number of rows to read at a time. Defaults to 50000. Defaults to None.
             path_to_file (str, optional): Path to local Excel file. Defaults to None.
-            url_to_file (str, optional): Link to a file on Sharepoint.
-                        (e.g : https://{tenant_name}.sharepoint.com/sites/{folder}/Shared%20Documents/Dashboard/file). Defaults to None.
             sheet_number (int, optional): Sheet number to be extracted from file. Counting from 0, if None all sheets are axtracted. Defaults to None.
             validate_excel_file (bool, optional): Check if columns in separate sheets are the same. Defaults to False.
             output_file_extension (str, optional): Output file extension - to allow selection of .csv for data which is not easy to handle with parquet. Defaults to ".csv".
@@ -61,6 +62,7 @@ class SharepointToADLS(Flow):
             adls_sp_credentials_secret (str, optional): The name of the Azure Key Vault secret containing a dictionary with
             ACCOUNT_NAME and Service Principal credentials (TENANT_ID, CLIENT_ID, CLIENT_SECRET) for the Azure Data Lake.
             Defaults to None.
+            overwrite_adls (bool, optional): Whether to overwrite files in the lake. Defaults to False.
             if_empty (str, optional): What to do if query returns no data. Defaults to "warn".
         """
         # SharepointToDF
@@ -73,6 +75,7 @@ class SharepointToADLS(Flow):
         self.validate_excel_file = validate_excel_file
 
         # AzureDataLakeUpload
+        self.overwrite = overwrite_adls
         self.adls_sp_credentials_secret = adls_sp_credentials_secret
         self.if_exists = if_exists
         self.output_file_extension = output_file_extension
@@ -136,6 +139,7 @@ class SharepointToADLS(Flow):
         file_to_adls_task.bind(
             from_path=self.local_file_path,
             to_path=self.adls_file_path,
+            overwrite=self.overwrite,
             sp_credentials_secret=self.adls_sp_credentials_secret,
             flow=self,
         )
@@ -146,6 +150,7 @@ class SharepointToADLS(Flow):
         json_to_adls_task.bind(
             from_path=self.local_json_path,
             to_path=self.adls_schema_file_dir_file,
+            overwrite=self.overwrite,
             sp_credentials_secret=self.adls_sp_credentials_secret,
             flow=self,
         )